requirements.txt
tweepy
nltk
transformers
pandas
matplotlib
python-dotenv
torch